refactor: 파일 구조 세팅
