+++ src/main/java/bridge/util/BridgeRandomNumberGenerator.java
@@ -1,4 +1,4 @@
-package bridge;
+package bridge.util;
 
 import camp.nextstep.edu.missionutils.Randoms;
